feat: 좋아요, 좋아요 취소 기능 구현

+++ src/main/java/com/swyp6/familytravel/feed/dto/FeedResponse.java
@@ -16,7 +16,7 @@ public class FeedResponse {
     private String content;
     private String place;
     private Long userId;
-    private Long likeCnt;
+    private Integer likeCnt;
     private LocalDate createDate;
     private List<String> imageList;
     private List<CommentResponse> commentList = new ArrayList<>();
